added suspensful animation to dealer's turn and refactored some css
